Better detect available memory in containers.
Use cgroup information if proc has it available.
